I must admin - I found the 'catchAll' behavio[u]r a little surprising
I tried updating your repo to zio 1.0.3 but had a bunch of errors, so this won't compile I don't think w/ your RC version.

Anyway, even on 1.0.3 'catchAll' has the same behaviour (I think the answer may be "works as designed").
'.sandbox' though works like .attempt in IO.

This area I think is tricky and will perhaps be whack-a-mole .... what would be surprising to one user wouldn't be to another.

I'm trying to picture this as a real-life PR review. Fundamentally I think having an unlawful 'URIO' (which I think is trying to represent the fact that you have a release operation which MUST succeed (somehow), because if you've not managed to do that then you have a resource leak and it's "blow the program up" time.

So, it seems that '.catchAll' is more for the equivalent of NonFatal errors, where '.sandbox' is more the equivalent of "I'm actually trying to recover from the non-typical exception, like OutOfMemoryException". In this case things like 'OutOfMemoryException' is from the JavaRuntime, where an unhandled close-block is treated as the equivalent 'fatal' error in the zio Runtime (In the zio code the BracketAccess results in a 'Zio.halt' ) 

That's how I view it anyway, though like you I was surprised to discover that by just looking at '.catchAll' in the code... I had to come to this understanding after playing with it, and this was a really good example.

So, if this WERE a PR for a project, although using '.sandbox' does have this run like the cats IO example, in reality I'd propose a fix of:
(1) communicating my new-found .sandbox vs .catchCall knowledge to the whole team/project and
(2) putting the fix in the close operation (rather than the '.sandbox' call), like:

```
val closeCorrect: AutoCloseable => URIO[Any, Any] = x => ZIO { x.close() }.ignore
```